feat: add separate basic storage file
